Move all tracing-related code to Tracers
No other changse have been made besides fixing up the imports.
