+++ cardano-node.cabal
@@ -69,6 +69,7 @@ executable cardano-node
                        GitRevFromGit
                        Paths_cardano_node
                        TraceAcceptor
+                       Tracers
                        TxSubmission
                        Run
                        Topology
